refactor: move file preview in seperate component
